module creation: #team_statable
